+++ projects/packages/search/src/dashboard/components/pages/dashboard-page.jsx
@@ -1,15 +1,16 @@
-import { JetpackFooter, JetpackLogo } from '@automattic/jetpack-components';
+import { JetpackFooter, JetpackSearchLogo, Button } from '@automattic/jetpack-components';
+import { useProductCheckoutWorkflow } from '@automattic/jetpack-connection';
 import { useSelect, useDispatch } from '@wordpress/data';
-import { createInterpolateElement } from '@wordpress/element';
-import { __, sprintf } from '@wordpress/i18n';
-import DonutMeterContainer from 'components/donut-meter-container';
+import { __ } from '@wordpress/i18n';
 import NoticesList from 'components/global-notices';
 import Loading from 'components/loading';
 import MockedSearch from 'components/mocked-search';
 import ModuleControl from 'components/module-control';
 import RecordMeter from 'components/record-meter';
-import React from 'react';
+import React, { useCallback } from 'react';
 import { STORE_ID } from 'store';
+import FirstRunSection from './sections/first-run-section';
+import PlanUsageSection from './sections/plan-usage-section';
 import './dashboard-page.scss';
 
 /**
@@ -24,6 +25,23 @@ export default function DashboardPage( { isLoading = false } ) {
 	useSelect( select => select( STORE_ID ).getSearchModuleStatus(), [] );
 	useSelect( select => select( STORE_ID ).getSearchStats(), [] );
 
+	const domain = useSelect( select => select( STORE_ID ).getCalypsoSlug() );
+	const siteAdminUrl = useSelect( select => select( STORE_ID ).getSiteAdminUrl() );
+
+	// Prepare Checkout action and loading status
+	const { fetchSearchPlanInfo } = useDispatch( STORE_ID );
+	const checkSiteHasSearchProduct = useCallback(
+		() => fetchSearchPlanInfo().then( response => response?.supports_search ),
+		[ fetchSearchPlanInfo ]
+	);
+	const { run: sendPaidPlanToCart, hasCheckoutStarted } = useProductCheckoutWorkflow( {
+		productSlug: 'jetpack_search',
+		redirectUrl: `${ siteAdminUrl }admin.php?page=jetpack-search`,
+		siteProductAvailabilityHandler: checkSiteHasSearchProduct,
+		from: 'jetpack-search',
+		siteSuffix: domain,
+	} );
+
 	const isPageLoading = useSelect(
 		select =>
 			select( STORE_ID ).isResolving( 'getSearchModuleStatus' ) ||
@@ -32,20 +50,28 @@ export default function DashboardPage( { isLoading = false } ) {
 			! select( STORE_ID ).hasStartedResolution( 'getSearchStats' ) ||
 			select( STORE_ID ).isResolving( 'getSearchPlanInfo' ) ||
 			! select( STORE_ID ).hasStartedResolution( 'getSearchPlanInfo' ) ||
-			isLoading,
-		[ isLoading ]
+			isLoading ||
+			hasCheckoutStarted,
+		[ isLoading, hasCheckoutStarted ]
 	);
 
 	// Introduce the gate for new pricing with URL parameter `new_pricing_202208=1`
 	const isNewPricing = useSelect( select => select( STORE_ID ).isNewPricing202208(), [] );
-	const siteAdminUrl = useSelect( select => select( STORE_ID ).getSiteAdminUrl() );
+	const isJustUpgraded = useSelect(
+		select => select( STORE_ID ).isFeatureEnabled( 'just_upgraded' ),
+		[]
+	);
 
+	const tierSlug = useSelect( select => select( STORE_ID ).getTierSlug() );
+
+	const isDisabledFromOverLimit = useSelect( select =>
+		select( STORE_ID ).getDisabledFromOverLimit()
+	);
 	const updateOptions = useDispatch( STORE_ID ).updateJetpackSettings;
 	const isInstantSearchPromotionActive = useSelect( select =>
 		select( STORE_ID ).isInstantSearchPromotionActive()
 	);
 
-	const domain = useSelect( select => select( STORE_ID ).getCalypsoSlug() );
 	const upgradeBillPeriod = useSelect( select => select( STORE_ID ).getUpgradeBillPeriod() );
 
 	const supportsOnlyClassicSearch = useSelect( select =>
@@ -72,29 +98,52 @@ export default function DashboardPage( { isLoading = false } ) {
 	const handleLocalNoticeDismissClick = useDispatch( STORE_ID ).removeNotice;
 	const notices = useSelect( select => select( STORE_ID ).getNotices(), [] );
 
+	// Plan Info data
+	const recordMeterInfo = {
+		lastIndexedDate,
+		postCount,
+		postTypeBreakdown,
+		postTypes,
+		tierMaximumRecords,
+	};
+
 	return (
 		<>
 			{ isPageLoading && <Loading /> }
 			{ ! isPageLoading && (
 				<div className="jp-search-dashboard-page">
-					<Header />
+					<Header
+						isUpgradable={ isNewPricing && ! tierSlug }
+						sendPaidPlanToCart={ sendPaidPlanToCart }
+					/>
 					<MockedSearchInterface
 						supportsInstantSearch={ supportsInstantSearch }
 						supportsOnlyClassicSearch={ supportsOnlyClassicSearch }
 					/>
-					{ isNewPricing && <MockUsageMeter /> }
-					<RecordMeter
-						postCount={ postCount }
-						postTypeBreakdown={ postTypeBreakdown }
-						tierMaximumRecords={ tierMaximumRecords }
-						lastIndexedDate={ lastIndexedDate }
-						postTypes={ postTypes }
-					/>
+					{ isNewPricing && (
+						<PlanInfo
+							hasIndex={ postCount !== 0 }
+							recordMeterInfo={ recordMeterInfo }
+							tierSlug={ tierSlug }
+							sendPaidPlanToCart={ sendPaidPlanToCart }
+							isJustUpgraded={ isJustUpgraded }
+						/>
+					) }
+					{ ! isNewPricing && (
+						<RecordMeter
+							postCount={ postCount }
+							postTypeBreakdown={ postTypeBreakdown }
+							tierMaximumRecords={ tierMaximumRecords }
+							lastIndexedDate={ lastIndexedDate }
+							postTypes={ postTypes }
+						/>
+					) }
 					<div className="jp-search-dashboard-bottom">
 						<ModuleControl
 							siteAdminUrl={ siteAdminUrl }
 							updateOptions={ updateOptions }
 							domain={ domain }
+							isDisabledFromOverLimit={ isDisabledFromOverLimit }
 							isInstantSearchPromotionActive={ isInstantSearchPromotionActive }
 							upgradeBillPeriod={ upgradeBillPeriod }
 							supportsOnlyClassicSearch={ supportsOnlyClassicSearch }
@@ -118,75 +167,42 @@ export default function DashboardPage( { isLoading = false } ) {
 	);
 }
 
-const MockUsageMeter = () => {
+const PlanInfo = ( {
+	hasIndex,
+	recordMeterInfo,
+	tierSlug,
+	sendPaidPlanToCart,
+	isJustUpgraded,
+} ) => {
+	// Site Info
+	// TODO: Investigate why this isn't returning anything useful.
+	const siteTitle = useSelect( select => select( STORE_ID ).getSiteTitle() ) || 'your site';
+	// Plan Info data
+	const currentPlan = useSelect( select => select( STORE_ID ).getCurrentPlan() );
+	const currentUsage = useSelect( select => select( STORE_ID ).getCurrentUsage() );
+	const latestMonthRequests = useSelect( select => select( STORE_ID ).getLatestMonthRequests() );
+	const planInfo = { currentPlan, currentUsage, latestMonthRequests, tierSlug };
+
 	return (
-		<div className="jp-search-dashboard-wrap jp-search-dashboard-meter-wrap">
-			<div className="jp-search-dashboard-row">
-				<div className="lg-col-span-2 md-col-span-1 sm-col-span-0"></div>
-				<div className="jp-search-dashboard-meter-wrap__content lg-col-span-8 md-col-span-6 sm-col-span-4">
-					<h2>
-						{ createInterpolateElement(
-							sprintf(
-								// translators: %1$s: usage period, %2$s: plan name
-								__( 'Your usage <s>%1$s (%2$s)</s>', 'jetpack-search-pkg' ),
-								'Sep 28-Oct 28',
-								__( 'Free plan', 'jetpack-search-pkg' )
-							),
-							{
-								s: <span />,
-							}
-						) }
-					</h2>
-					<div className="usage-meter-group">
-						<DonutMeterContainer
-							title={ __( 'Site records', 'jetpack-search-pkg' ) }
-							current={ 1250 }
-							limit={ 5000 }
-						/>
-						<DonutMeterContainer
-							title={ __( 'Search requests', 'jetpack-search-pkg' ) }
-							current={ 125 }
-							limit={ 500 }
-						/>
-					</div>
-					<div className="upgrade-trigger">
-						<div>
-							{ __(
-								'Do you want to increase your site records and search requests?',
-								'jetpack-search-pkg'
-							) }
-						</div>
-						<strong>
-							{ __( 'Upgrade now and avoid any future interruption!', 'jetpack-search-pkg' ) }
-						</strong>
-						<svg
-							width="19"
-							height="15"
-							viewBox="0 0 19 15"
-							fill="none"
-							xmlns="http://www.w3.org/2000/svg"
-						>
-							<path
-								d="M18.2969 7.50391C18.2969 7.11719 18.1328 6.71875 17.8633 6.44922L11.9102 0.519531C11.6055 0.214844 11.2539 0.0625 10.9141 0.0625C10.0703 0.0625 9.49609 0.648438 9.49609 1.42188C9.49609 1.85547 9.68359 2.18359 9.95312 2.45312L11.8164 4.32812L13.8555 6.19141L11.8516 6.07422H1.83203C0.941406 6.07422 0.34375 6.64844 0.34375 7.50391C0.34375 8.35938 0.941406 8.94531 1.83203 8.94531H11.8516L13.8555 8.82812L11.8164 10.6914L9.95312 12.5664C9.68359 12.8242 9.49609 13.1641 9.49609 13.5859C9.49609 14.3594 10.0703 14.9453 10.9141 14.9453C11.2539 14.9453 11.6055 14.793 11.9102 14.5L17.8633 8.57031C18.1328 8.30078 18.2969 7.90234 18.2969 7.50391Z"
-								fill="#069E08"
-							/>
-						</svg>
-					</div>
-					<div className="usage-meter-about">
-						{ createInterpolateElement(
-							__(
-								'Tell me more about <u>record indexing and request limits</u>',
-								'jetpack-search-pkg'
-							),
-							{
-								u: <u />,
-							}
-						) }
-					</div>
-				</div>
-				<div className="lg-col-span-2 md-col-span-1 sm-col-span-0"></div>
-			</div>
-		</div>
+		<>
+			{ ! hasIndex && <FirstRunSection siteTitle={ siteTitle } planInfo={ planInfo } /> }
+			{ hasIndex && (
+				<>
+					<PlanUsageSection
+						planInfo={ planInfo }
+						sendPaidPlanToCart={ sendPaidPlanToCart }
+						isJustUpgraded={ isJustUpgraded }
+					/>
+					<RecordMeter
+						postCount={ recordMeterInfo.postCount }
+						postTypeBreakdown={ recordMeterInfo.postTypeBreakdown }
+						tierMaximumRecords={ recordMeterInfo.tierMaximumRecords }
+						lastIndexedDate={ recordMeterInfo.lastIndexedDate }
+						postTypes={ recordMeterInfo.postTypes }
+					/>
+				</>
+			) }
+		</>
 	);
 };
 
@@ -233,13 +249,20 @@ const Footer = () => {
 	);
 };
 
-const Header = () => {
+const Header = ( { isUpgradable, sendPaidPlanToCart } ) => {
+	const buttonLinkArgs = {
+		children: __( 'Upgrade Jetpack Search', 'jetpack-search-pkg' ),
+		variant: 'link',
+		onClick: sendPaidPlanToCart,
+	};
+
 	return (
 		<div className="jp-search-dashboard-header jp-search-dashboard-wrap">
 			<div className="jp-search-dashboard-row">
 				<div className="lg-col-span-12 md-col-span-8 sm-col-span-4">
 					<div className="jp-search-dashboard-header__logo-container">
-						<JetpackLogo className="jp-search-dashboard-header__masthead" />
+						<JetpackSearchLogo className="jp-search-dashboard-header__masthead" />
+						{ isUpgradable && <Button { ...buttonLinkArgs } /> }
 					</div>
 				</div>
 			</div>