+++ projects/packages/videopress/src/client/admin/components/pricing-section/index.tsx
@@ -21,10 +21,15 @@ const PricingPage = () => {
 	const { siteSuffix, adminUri } = window.jetpackVideoPressInitialState;
 	const { siteProduct, product } = usePlan();
 	const { pricingForUi } = siteProduct;
-	const { handleRegisterSite, userIsConnecting } = useConnection( { redirectUri: adminUri } );
+	const { registrationNonce } = window.jetpackVideoPressInitialState;
+	const { handleRegisterSite, userIsConnecting } = useConnection( {
+		redirectUri: adminUri,
+		from: 'jetpack-videopress',
+		registrationNonce,
+	} );
 	const [ isConnecting, setIsConnection ] = useState( false );
 
-	const { run } = useProductCheckoutWorkflow( {
+	const { run, hasCheckoutStarted } = useProductCheckoutWorkflow( {
 		siteSuffix,
 		productSlug: product.productSlug,
 		redirectUrl: adminUri,
@@ -50,10 +55,15 @@ const PricingPage = () => {
 						price={ pricingForUi.fullPrice }
 						offPrice={ pricingForUi.discountPrice }
 						promoLabel={ __( '50% off', 'jetpack-videopress-pkg' ) }
-						leyend={ __( '/month, billed yearly', 'jetpack-videopress-pkg' ) }
+						legend={ __( '/month, billed yearly', 'jetpack-videopress-pkg' ) }
 						currency={ pricingForUi.currencyCode }
 					/>
-					<Button onClick={ run } fullWidth disabled={ isConnecting }>
+					<Button
+						onClick={ run }
+						isLoading={ hasCheckoutStarted }
+						fullWidth
+						disabled={ isConnecting || hasCheckoutStarted }
+					>
 						{ __( 'Get VideoPress', 'jetpack-videopress-pkg' ) }
 					</Button>
 				</PricingTableHeader>
@@ -64,7 +74,12 @@ const PricingPage = () => {
 			</PricingTableColumn>
 			<PricingTableColumn>
 				<PricingTableHeader>
-					<ProductPrice price={ 0 } leyend="" currency="USD" hidePriceFraction />
+					<ProductPrice
+						price={ 0 }
+						legend=""
+						currency={ pricingForUi.currencyCode }
+						hidePriceFraction
+					/>
 					<Button
 						fullWidth
 						variant="secondary"
@@ -72,8 +87,8 @@ const PricingPage = () => {
 							setIsConnection( true );
 							handleRegisterSite();
 						} }
-						isLoading={ userIsConnecting || isConnecting }
-						disabled={ userIsConnecting || isConnecting }
+						isLoading={ userIsConnecting }
+						disabled={ userIsConnecting || isConnecting || hasCheckoutStarted }
 					>
 						{ __( 'Start for free', 'jetpack-videopress-pkg' ) }
 					</Button>