+++ projects/plugins/jetpack/extensions/blocks/videopress/edit.js
@@ -30,7 +30,7 @@ import { useSelect, withDispatch, withSelect } from '@wordpress/data';
 import { Component, createRef, Fragment } from '@wordpress/element';
 import { escapeHTML } from '@wordpress/escape-html';
 import { __, _x, sprintf } from '@wordpress/i18n';
-import { Icon, pencil } from '@wordpress/icons';
+import { Icon } from '@wordpress/icons';
 import classnames from 'classnames';
 import { get, indexOf } from 'lodash';
 import { useEffect } from 'react';
@@ -494,9 +494,8 @@ const VideoPressEdit = CoreVideoEdit =>
 						<ToolbarGroup>
 							<ToolbarButton
 								className="components-icon-button components-toolbar__control"
-								label={ __( 'Edit video', 'jetpack' ) }
+								text={ __( 'Replace', 'jetpack' ) }
 								onClick={ this.switchToEditing }
-								icon={ <Icon icon={ pencil } /> }
 							/>
 						</ToolbarGroup>
 					</BlockControls>